+++ src/dtflw/storage/azure.py
@@ -43,10 +43,10 @@ class AzureStorage(FileStorageBase):
         """
         Returns the base path.
         """
-        return f"wasbs://{self.__container_name}@{self.__account_name}.blob.core.windows.net"
+        return f"wasbs://{self.__container_name}@{self.__account_name}.blob.core.windows.net/"
 
 
-def init_storage(account_name: str, container_name: str, root_dir: str = None, spark: SparkSession = None, dbutils=None):
+def init_storage(account_name: str, container_name: str, root_dir: str = None, spark: SparkSession = None, dbutils=None) -> AzureStorage:
     """
     Returns a new instance of AzureStorage. 
     It is suggested using this factory function instead of the constructor of AzureStorage class.